Move utility functions into util module
